Time: 31 ms (14.62%), Space: 23 MB (10.17%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -18,7 +18,7 @@ public:
         if (root != nullptr) q.push(root);
         while (!q.empty()) {
             int size = q.size();
-            vector<double> level;            
+            vector<long> level;            
             for (int i = 0; i < size; i++) {
                 cur = q.front();
                 q.pop();
